feat(casino): unify game reads and edits in GamingService
Fold the GameCatalogService composer back into GamingService so games
have one owner: a single paginated, filterable listGames serves both
the public catalog and the backoffice list, and updateGame lives next
to the reads it returns. Entity services stay split by entity.

- public GET /gaming/games becomes paginated with q/provider/category
  filters, always scoped to active games
- GET /backoffice/gaming/games shares the same query with an optional
  isActive filter
- updateGame and GameSlugTakenError move to GamingService
- drop GameCatalogService; its tests merge into gaming.service tests
